FIXED BUGS

+++ wtgallery/views.py
@@ -3,7 +3,7 @@
 import os
 from django.contrib.auth.models import User, AnonymousUser
 from django.core.mail import send_mail
-from django.http import JsonResponse, HttpResponse, HttpResponseRedirect, FileResponse, Http404
+from django.http import JsonResponse, HttpResponse, HttpResponseRedirect, Http404, FileResponse
 from django.contrib.auth import login, authenticate
 from django.contrib.auth.decorators import login_required
 from django.db.models import Q
@@ -14,12 +14,18 @@ from django.views.decorators.csrf import csrf_exempt
 from django.views.generic import ListView
 from django.core.paginator import Paginator
 from django.core.files import File
+import boto3
 
 from itertools import chain
 from pixzium import settings
 from .forms import ImageForm, SignUpForm, LoginForm, VideoForm, MusicForm
 from django.contrib import messages
 from .models import Image, Video, Music, Profile
+
+resource = boto3.resource('s3',
+                          aws_access_key_id=settings.AWS_ACCESS_KEY_ID,
+                          aws_secret_access_key=settings.AWS_SECRET_ACCESS_KEY,
+                          )
 
 
 def index(request):
@@ -45,7 +51,7 @@ def index(request):
     portfolio_list = Image.objects.all().order_by('-uploaded_at').filter(status__exact='A')
 
     current_page = request.GET.get('page', 1)
-    paginator = Paginator(portfolio_list, 4)
+    paginator = Paginator(portfolio_list, 12)
     try:
         portfolio = paginator.page(current_page)
     except PageNotAnInteger:
@@ -70,7 +76,6 @@ def my_account(request):
     current_user = request.user
     profileRecord = Profile.objects.get_or_create(user=current_user)[0]
     image = Image.objects.filter(user__user__username__exact=profileRecord)
-    print(image)
     video = Video.objects.filter(user__user__username__exact=profileRecord)
     music = Music.objects.filter(user__user__username__exact=profileRecord)
 
@@ -83,27 +88,31 @@ def my_account(request):
             messages.success(request, 'Message Submitted Successfully. Our Team will reply you ASAP.')
         # User Profile Update(Basic Info)
         if 'btnSave' in request.POST:
-            FirstName = request.POST.get("firstname", None)
-            LastName = request.POST.get("lastname", None)
-            Mobile = request.POST.get("mobile", 000)
-            Address = request.POST.get("address", None)
-            City = request.POST.get("city", None)
-            State = request.POST.get("state", None)
-            Country = request.POST.get("country", None)
-            Pincode = request.POST.get("pincode", 000)
-
+            FirstName = request.POST.get("firstname", profileRecord.user.first_name)
+            LastName = request.POST.get("lastname", profileRecord.user.last_name)
+            Mobile = request.POST.get("mobile", profileRecord.mobile)
+            Address = request.POST.get("address", profileRecord.address)
+            City = request.POST.get("city", profileRecord.city)
+            State = request.POST.get("state", profileRecord.state)
+            Country = request.POST.get("country", profileRecord.country)
+            Pincode = request.POST.get("pincode", profileRecord.pincode)
+            Email = request.POST.get('email', current_user.email)
             current_user.first_name = FirstName
             current_user.last_name = LastName
-            profileRecord.mobile = Mobile
+            current_user.email = Email
             profileRecord.address = Address
             profileRecord.city = City
             profileRecord.state = State
             profileRecord.country = Country
-            profileRecord.pincode = Pincode
-
-            profileRecord.save()
-            current_user.save()
-            messages.success(request, 'Your Profile is successfully Uploaded')
+            if Mobile[-10:].isdigit():
+                profileRecord.mobile = Mobile
+                if Pincode.isdigit():
+                    profileRecord.pincode = Pincode
+                profileRecord.save()
+                current_user.save()
+                messages.success(request, 'Your Profile is successfully Uploaded')
+            else:
+                messages.warning(request, 'Enter Fields properly. Mobile should be digits only!')
         # User Public Profile Update
         if 'btnPPSave' in request.POST:
             # user_id = request.POST.get("user_id", None)
@@ -151,7 +160,7 @@ def my_account(request):
                'image': image,
                'video': video,
                'music': music,
-               'followers': '',}
+               'followers': '', }
     return render(request, "dashboard.html", context)
 
 
@@ -162,35 +171,36 @@ def dashboard(request):
 
 @login_required
 def approval(request):
-    if request.method == 'POST':
-        obj_id = int(request.POST.get("obj_id"))
-        status = str(request.POST.get("status"))
-        user_id = request.POST.get("user_id")
-        type = request.POST.get("type")
-        print(user_id)
-        user = User.objects.get(id=user_id)
-        u_email = user.email
+    if request.user.is_superuser:
+        if request.method == 'POST':
+            obj_id = int(request.POST.get("obj_id"))
+            status = str(request.POST.get("status"))
+            user_id = request.POST.get("user_id")
+            type = request.POST.get("type")
+            user = User.objects.get(id=user_id)
+            u_email = user.email
 
-        if type == 'image':
-            sts = Image.objects.get(id=obj_id)
-        elif type == 'video':
-            sts = Video.objects.get(id=obj_id)
-        else:
-            sts = Music.objects.get(id=obj_id)
+            if type == 'image':
+                sts = Image.objects.get(id=obj_id)
+            elif type == 'video':
+                sts = Video.objects.get(id=obj_id)
+            else:
+                sts = Music.objects.get(id=obj_id)
 
-        if status == "A":
-            sts.status = status
-            sts.save()
-        else:
-            subject = 'Warning Message for Uploaded Image from Pixzium'
-            message = f'Hi {user}, we have found Inappropriate Image. This is a waring Email, Please Follow ' \
-                      f'Our Company Guideline. '
-            email_from = settings.EMAIL_HOST_USER
-            recipient_list = [u_email, ]
-            send_mail(subject, message, email_from, recipient_list)
-            sts.delete()
-            messages.success(request, 'Record Deleted Successfully.')
-
+            if status == "A":
+                sts.status = status
+                sts.save()
+            else:
+                subject = 'Warning Message for Uploaded Image from Pixzium'
+                message = f'Hi {user}, we have found Inappropriate Image. This is a waring Email, Please Follow ' \
+                          f'Our Company Guideline. '
+                email_from = settings.EMAIL_HOST_USER
+                recipient_list = [u_email, ]
+                send_mail(subject, message, email_from, recipient_list)
+                sts.delete()
+                messages.success(request, 'Record Deleted Successfully.')
+    else:
+        raise Http404
     images = Image.objects.all().filter(status__exact='P')
     videos = Video.objects.all().filter(status__exact='P')
     music = Music.objects.all().filter(status__exact='P')
@@ -269,10 +279,9 @@ def profile(request, username):
     return render(request, 'profile.html', context)
 
 
-def photo_detail(request, id):
-    image = get_object_or_404(Image, id=id)
+def photo_detail(request, slug):
+    image = get_object_or_404(Image, slug=slug)
     profile = Profile.objects.filter(user=image.user.id).first()
-    print(profile)
 
     tagsList = []
     for tag in image.tags.all():
@@ -288,7 +297,7 @@ def photo_detail(request, id):
 def image(request):
     portfolio_list = Image.objects.all().order_by('-uploaded_at').filter(status__exact='A')
     current_page = request.GET.get('page', 1)
-    paginator = Paginator(portfolio_list, 4)
+    paginator = Paginator(portfolio_list, 12)
     try:
         portfolio = paginator.page(current_page)
     except PageNotAnInteger:
@@ -303,7 +312,7 @@ def image(request):
 def video(request):
     video_list = Video.objects.all().order_by('-uploaded_at').filter(status__exact='A')
     current_page = request.GET.get('page', 1)
-    paginator = Paginator(video_list, 4)
+    paginator = Paginator(video_list, 12)
     try:
         videos = paginator.page(current_page)
     except PageNotAnInteger:
@@ -317,7 +326,6 @@ def video(request):
 
 def music(request):
     portfolio = Music.objects.all().order_by('-uploaded_at')
-    print(portfolio)
     context = {"portfolio": portfolio}
 
     return render(request, "music.html", context)
@@ -332,8 +340,8 @@ def upload(request):
 
         ext = os.path.splitext(str(request.FILES['file']))[-1].lower()
 
-        if ext == ".mp4":
-            print("mp4 file!")
+        if ext == ".mp4" or ext == ".mkv" or ext == ".mov":
+            print(f"{ext} file!")
 
             if video.is_valid:
                 # form.save()
@@ -345,34 +353,8 @@ def upload(request):
 
                 return redirect('upload')
 
-        elif ext == ".mkv":
-            print("mkv file!")
-
-            if video.is_valid:
-                # form.save()
-                fs = video.save(commit=False)
-                fs.user = Profile.objects.get(user=request.user)
-                fs.save()
-                video.save_m2m()
-                messages.success(request, 'Video successfully Uploaded, It will be Visible after reviewed by our Team.')
-
-                return redirect('upload')
-
-        elif ext == ".mov":
-            print("mov file!")
-
-            if video.is_valid:
-                # form.save()
-                fs = video.save(commit=False)
-                fs.user = Profile.objects.get(user=request.user)
-                fs.save()
-                video.save_m2m()
-                messages.success(request, 'Video successfully Uploaded, It will be Visible after reviewed by our Team.')
-
-                return redirect('upload')
-
-        elif ext == ".mp3":
-            print("mp3 file!")
+        elif ext == ".mp3" or ext == ".avi" or ext == ".m4a":
+            print(f"{ext} file!")
 
             if music.is_valid:
                 # form.save()
@@ -384,14 +366,10 @@ def upload(request):
 
                 return redirect('upload')
 
-        elif ext == ".png" or ".jpg" or ".jpeg":
-            print("image file!")
+        elif ext == ".png" or ext == ".jpg" or ext == ".jpeg":
+            print(f"{ext} file!")
 
             if image.is_valid():
-                # if nude.is_nude(request.FILES['file']):
-                #     messages.warning(request, 'Inappropriate image detected, This is against our company policy !!')
-                # else:
-                # form.save()
                 fs = image.save(commit=False)
                 fs.user = Profile.objects.get(user=request.user)
                 fs.save()
@@ -474,15 +452,20 @@ def music_views(request):
         obj = Music.objects.get(pk=pk)
         obj.views += 1
         obj.save()
-
-    return JsonResponse({'status': obj.views})
+        return JsonResponse({'status': obj.views})
+    raise Http404
 
 
 @login_required
 def count_likes(request):
     if request.method == 'GET':
         id = request.GET.get('post_id')
-        post = get_object_or_404(Image, id=id)
+        images = Image.objects.all()
+        videos = Video.objects.all()
+        musics = Music.objects.all()
+        object_list = list(chain(images, videos, musics))
+
+        post = get_object_or_404(object_list, id=id)
 
         liked = False
         if post.likes.filter(id=request.user.id).exists():
@@ -532,13 +515,21 @@ def count_downloads(request, type, slug):
             obj = Music.objects.get(slug=slug)
         elif type == 'video':
             obj = Video.objects.get(slug=slug)
+            return HttpResponse(obj.file.url, content_type='text/plain')
         else:
-            return Http404
+            print('THIS ERROR FOR TYPE')
+            raise Http404
         obj.total_downloads += 1
-        file = open(f'.{obj.file.url}', 'rb')
-        obj.save()
-        return FileResponse(file, content_type='application/force-download')
-    return Http404
+        try:
+            # TODO: Works absolutely fine with images and music but not videos
+            filename = obj.file.name.split('/')[-1]
+            response = HttpResponse(obj.file, content_type='text/plain')
+            response['Content-Disposition'] = f'attachment; filename={filename}'
+            obj.save()
+            return response
+        except:
+            raise Http404
+    raise Http404
 
 
 def sitemap(request):
@@ -551,8 +542,6 @@ def subscription(request):
 
 def about(request):
     return render(request, "about.html")
-
-
 
 # class IndexView(generic.ListView):
 #     model = Image
